+++ Tests/AdvancedMappingCrud.RichDomain.ServiceModel.Tests/AdvancedMappingCrud.RichDomain.ServiceModel.Tests.Domain/Services/CategoriesService.cs
@@ -32,5 +32,12 @@ namespace AdvancedMappingCrud.RichDomain.ServiceModel.Tests.Domain.Services
             // TODO: Implement DoIt (CategoriesService) functionality
             throw new NotImplementedException("Implement your domain service logic here...");
         }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
+        public async Task ManualAsync(CancellationToken cancellationToken = default)
+        {
+            // TODO: Implement ManualAsync (CategoriesService) functionality
+            throw new NotImplementedException("Implement your domain service logic here...");
+        }
     }
 }